Refactor the runtime: move content working group code to the separate file.
