+++ runtime/src/integration/mod.rs
@@ -1,3 +1,4 @@
+pub mod content_working_group;
 pub mod proposals;
 pub mod storage;
 pub mod working_group;
